net: stream: add a new option to automatically reconnect
RH-Author: Laurent Vivier <lvivier@redhat.com>
RH-MergeRequest: 148: net: stream: add a new option to automatically reconnect
RH-Bugzilla: 2169232
RH-Acked-by: Eugenio Pérez <eperezma@redhat.com>
RH-Acked-by: Cindy Lu <lulu@redhat.com>
RH-Acked-by: MST <mst@redhat.com>
RH-Acked-by: David Gibson (Red Hat) <dgibson@redhat.com>
RH-Commit: [2/2] 9b87647a9ed2e7c1b91bdfa9d0a736e091c892a5 (lvivier/qemu-kvm-centos)

In stream mode, if the server shuts down there is currently
no way to reconnect the client to a new server without removing
the NIC device and the netdev backend (or to reboot).

This patch introduces a reconnect option that specifies a delay
to try to reconnect with the same parameters.

Add a new test in qtest to test the reconnect option and the
connect/disconnect events.

Signed-off-by: Laurent Vivier <lvivier@redhat.com>
Signed-off-by: Jason Wang <jasowang@redhat.com>
(cherry picked from commit b95c0d4440950fba6dbef0f781962911fa42abdb)

Patch-name: kvm-net-stream-add-a-new-option-to-automatically-reconne.patch
Patch-id: 107
Patch-present-in-specfile: True