spec: Disable VDUSE
RH-Author: Kevin Wolf <kwolf@redhat.com>
RH-MergeRequest: 142: spec: Disable VDUSE
RH-Bugzilla: 2128222
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Hanna Czenczek <hreitz@redhat.com>
RH-Commit: [1/1] 72f575d0d570a9a058cef89193f5fb8f9fa6a653 (kmwolf/centos-qemu-kvm)

With the recent rebases, we inherited VDUSE support from upstream and
configure automatically enabled it. We don't actually want to support it
at this point, and in fact it doesn't even work with the current kernel.
All it does is making --help show these VDUSE options that don't work.

Therefore, disable the VDUSE related options in the build.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>