virtio-rng-pci: fix transitional migration compat for vectors
RH-Author: Dr. David Alan Gilbert <dgilbert@redhat.com>
RH-MergeRequest: 147: virtio-rng-pci: fix transitional migration compat for vectors
RH-Bugzilla: 2162569
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Commit: [1/1] 6e2bd111cd56808fccf2c0464a40f7784fd893a2 (dagrh/c-9-s-qemu-kvm)

In upstream bad9c5a5166/downstream 46e08bafe9ed  I fixed the virito-rng-pci
migration compatibility, but it was discovered that we also need to fix
the other aliases of the device for the transitional cases.

I've sent upstream:
https://lists.gnu.org/archive/html/qemu-devel/2023-02/msg01926.html
but downstream we need to change the downstream machine type anyway,
so it's not quite identical.

Fixes: 9ea02e8f1 ('virtio-rng-pci: Allow setting nvectors, so we can use MSI-X')

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>

Patch-name: kvm-virtio-rng-pci-fix-transitional-migration-compat-for.patch
Patch-id: 105
Patch-present-in-specfile: True