multifd: Fix the number of channels ready
RH-Author: Leonardo Brás <leobras@redhat.com>
RH-MergeRequest: 171: multifd: Fix the number of channels ready
RH-Bugzilla: 2196289
RH-Acked-by: Peter Xu <peterx@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>
RH-Commit: [1/1] a5e271ba249d85b27a68d3cff10480ca3a112c5d (LeoBras/centos-qemu-kvm)

We don't wait in the sem when we are doing a sync_main.  Make it wait
there.  To make things clearer, we mark the channel ready at the
begining of the thread loop.

Signed-off-by: Juan Quintela <quintela@redhat.com>
Reviewed-by: Fabiano Rosas <farosas@suse.de>
(cherry picked from commit d2026ee117147893f8d80f060cede6d872ecbd7f)
Signed-off-by: Leonardo Bras <leobras@redhat.com>

Patch-name: kvm-multifd-Fix-the-number-of-channels-ready.patch
Patch-id: 98
Patch-present-in-specfile: True